upd HashdeepLine support different output formats; fix truncation
hashdeep's default output is
size,md5,sha256,filename

but I found some older checksum files that are just
md5,filename

so in order to accommodate that with the same logic, default to checking
on md5.

in some instances of running, somehow the program seemed to be
terminating before all comparisons were done, and the early termination
happened at different points of comparison. From the code alone, the
reason is not obvious to me, and I don't yet have acccess to GPT-4 to
ask why.

But it appears that adding a return Result in
```
fn compare_hashdeep_hashmaps(base_map: &HashMap<String, HashdeepLine>,
comp_map: &HashMap<String, HashdeepLine>) -> Result<CompareStats,
String>
```

causes the program to behave correctly and finish all comparisons.